pyproject.toml
[project]
name = "llm-parameter-guesser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-speech>=2.36.0",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "scipy>=1.17.0",
    "sounddevice>=0.5.5",
]
